clean and fix tests
