+++ fhirpipe/load/fhirstore.py
@@ -45,44 +45,3 @@ def save_many(instances):
     for instance in instances:
         store.create(instance)
         instances.refresh()
-
-
-fhir_ids = {}
-
-
-def find_fhir_resource(uri, identifier):
-    """
-    Return the first FHIR instance of some resource_type where the
-    identifier matches some identifier, if any is found.
-
-    args:
-        resource_type (str): name of the Fhir resource to search for
-        identifier (str): the provided id which could be an identifier of
-            some instance of the resource_type
-
-    return:
-        the fhir id of the instance is there is one found of resource_type
-        which has an identifier matching the provided identifier,
-        else None
-    """
-    db_client = get_mongo_client()[fhirpipe.global_config["fhirstore"]["database"]]
-
-    # NOTE For now, URI simply consists in the resource name, need discussion about this
-    resource_type = uri
-
-    if resource_type not in fhir_ids:
-        # If URI not found, don't do anything
-        if resource_type not in db_client.list_collection_names():
-            return
-
-        results = db_client[resource_type].find(
-            {"identifier": {"$elemMatch": {"value": {"$exists": True}}}},
-            ["id", "identifier"],
-        )
-
-        fhir_ids[resource_type] = {}
-        for r in results:
-            for r_identifier in r["identifier"]:
-                fhir_ids[resource_type][r_identifier["value"]] = r["id"]
-
-    return fhir_ids[resource_type].get(identifier)
