+++ test/unit/load/test_fhirstore.py
@@ -1,8 +1,6 @@
 from unittest import mock
 
 import fhirpipe.load.fhirstore as fhirstore
-from test.unit import mock_config
-from test.unit.load import mock_mongo_client
 
 
 @mock.patch("fhirpipe.load.fhirstore.get_fhirstore", return_value=mock.Mock())
@@ -25,16 +23,3 @@ def test_save_many(_):
             mock.call({"name": "instance3", "value": 3}),
         ]
     )
-
-
-@mock.patch("fhirpipe.load.fhirstore.get_mongo_client", return_value=mock_mongo_client)
-@mock.patch("fhirpipe.load.fhirstore.fhirpipe.global_config", mock_config)
-def test_find_fhir_resource(*_):
-
-    assert fhirstore.find_fhir_resource("Patient", "0001") == "123456"
-
-    # Use URI that doesn't exist:
-    assert fhirstore.find_fhir_resource("Observation", "0001") is None
-
-    # With identifier value that doesn't exist
-    assert fhirstore.find_fhir_resource("Observation", "0000") is None
